requirements.txt
yolov5 == 6.2.3
Pillow == 9.3.0
opencv-python == 4.6.0.66